requirements.txt
Django==3.2.6
django-rest-swagger==2.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-apscheduler
drf-yasg==1.20.0
PyJWT==2.4.0
mysqlclient==2.0.3
